requirements.txt
streamlit
pandas
requests
python-dotenv
pytz
supabase
xlsxwriter
openpyxl